requirements.txt
neuralprophet
xlrd
seaborn
fastapi
requests
python-dotenv
mysql-connector-python
numpy
pandas
plotly
yfinance
